Land the E/S/C wave work that was never committed
K3dev.md records E1-E6, S1/S5, C1-C6 and the Rust A-wave as complete, but the
tree carried them uncommitted, so the repository did not match its own tracker.
This lands them as one commit rather than folding them into unrelated work.

Includes: ler.py (Wilson CIs, estimate_ler, threshold sweeps, memory experiment),
bench.py, codes.py CSS logicals, dem.py / sinter_compat.py / pymatching_compat.py
DEM path, gpu_backend.py cupy probe, and the accompanying test modules
(test_ler, test_logical_operators, test_reference_accuracy, test_regression_gates,
test_dem_fastpath, test_gpu_probe_no_cupy) plus the scripts/ benchmark harness.

Also drops benchmark_results/ from tracking to match .gitignore -- the canonical
archive is Zenodo (doi.org/10.5281/zenodo.21339300), not the repo.

Not reviewed line-by-line in this session; landed because the full suite is green
across the whole tree (1237 passed, 1 skipped, 0 failed).

+++ python/tests/test_pymatching_parity_p_sweep.py
@@ -73,5 +73,5 @@ def _assert_parity(d, shots, seed, basis, noise):
 @pytest.mark.parametrize("d", [5, 7])
 def test_pymatching_parity_p_sweep(d, noise):
     """QECTOR/PyMatching LER parity at each physical error rate."""
-    seed = 20260625 + int(round(noise * 1000)) + d
+    seed = 20260625 + round(noise * 1000) + d
     _assert_parity(d, shots=3000, seed=seed, basis="x", noise=noise)
